# Conflicts:
#   src/main/resources/static/dodaj-przepis.js

+++ src/main/java/pw/paint/controller/RecipeController.java
@@ -10,12 +10,14 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 import pw.paint.DTOs.model.RecipeDto;
 import pw.paint.DTOs.model.ShortRecipeDto;
 import pw.paint.DTOs.requests.NewRecipeRequest;
 import pw.paint.DTOs.requests.SearchRequest;
 import pw.paint.exception.RecipeNotFoundException;
 import pw.paint.exception.UserNotFoundException;
+import pw.paint.service.JwtService;
 import pw.paint.service.RecipeService;
 
 import java.net.URI;
@@ -24,23 +26,64 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/recipe")
 public class RecipeController {
     private final RecipeService recipeService;
+    private final JwtService jwtService;
 
     @GetMapping("/tags")
     public ResponseEntity<List<String>> getAllTags() {
         return ResponseEntity.ok(recipeService.getAllTags());
     }
 
+//    @PostMapping("/new")
+//    public ResponseEntity<Void> createNewRecipe(@RequestParam(value = "image", required = false) MultipartFile image, @ModelAttribute NewRecipeRequest newRecipeRequest) {
+//        try {
+//            byte[] imageBytes = (image != null) ? image.getBytes() : null;
+//            return ResponseEntity.created(new URI("/recipe/" +
+//                    recipeService.createNewRecipe(newRecipeRequest, imageBytes))).build();
+//        } catch (UserNotFoundException ex) {
+//            return ResponseEntity.status(HttpStatus.CONFLICT)
+//                    .header("Error-message", ex.getMessage())
+//                    .build();
+//        } catch (Exception ex) {
+//            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+//                    .header("Error-message", ex.getMessage())
+//                    .build();
+//        }
+//    }
+
+//    @PostMapping("/new")
+//    public ResponseEntity<Void> createNewRecipe(@ModelAttribute NewRecipeRequest newRecipeRequest) {
+//        try {
+//            return ResponseEntity.created(new URI("/recipe/" +
+//                    recipeService.createNewRecipe(newRecipeRequest))).build();
+//        } catch (UserNotFoundException ex) {
+//            return ResponseEntity.status(HttpStatus.CONFLICT)
+//                    .header("Error-message", ex.getMessage())
+//                    .build();
+//        } catch (Exception ex) {
+//            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+//                    .header("Error-message", ex.getMessage())
+//                    .build();
+//        }
+//    }
+
     @PostMapping("/new")
-    public ResponseEntity<Void> createNewRecipe(@RequestBody NewRecipeRequest newRecipeRequest) {
+    public ResponseEntity<String> createNewRecipe(@RequestBody NewRecipeRequest newRecipeRequest,
+                                                @CookieValue(name = "token", required = false) String token ) {
         try {
-            return ResponseEntity.created(new URI("/recipe/" +
-                    recipeService.createNewRecipe(newRecipeRequest))).build();
+            if(token == null || !newRecipeRequest.getAuthor().equals(jwtService.extractUsername(token))){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                        .header("Error-message", "Forbidden")
+                        .build();
+            }
+            return ResponseEntity.status(HttpStatus.CREATED)
+                    .body(recipeService.createNewRecipe(newRecipeRequest).toString());
         } catch (UserNotFoundException ex) {
             return ResponseEntity.status(HttpStatus.CONFLICT)
                     .header("Error-message", ex.getMessage())
@@ -52,27 +95,41 @@ public class RecipeController {
         }
     }
 
-    @GetMapping("/{id}")
-    public ResponseEntity<RecipeDto> getRecipeById(@PathVariable String id) {
+    @PutMapping("/set-img/{id}")
+    public ResponseEntity<Void> setImage(@PathVariable String id, @ModelAttribute MultipartFile image,
+                                         @CookieValue(name = "token", required = false) String token ) {
         try {
-            return ResponseEntity.ok(recipeService.getRecipeById(new ObjectId(id)));
+            if(token == null){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                        .header("Error-message", "Forbidden")
+                        .build();
+            }
+            recipeService.setImage(id, image);
+            return ResponseEntity.status(HttpStatus.CREATED).build();
         } catch (RecipeNotFoundException ex) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .header("Error-Message", ex.getMessage())
+            return ResponseEntity.status(HttpStatus.CONFLICT)
+                    .header("Error-message", ex.getMessage())
+                    .build();
+        } catch (Exception ex) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .header("Error-message", ex.getMessage())
                     .build();
         }
     }
 
-    @GetMapping("/image/{id}")
-    public ResponseEntity<byte[]> getImage(@PathVariable String id) {
+    @GetMapping("/{id}")
+    public ResponseEntity<RecipeDto> getRecipeById(@PathVariable String id,
+                                                   @CookieValue(name = "token", required = false) String token ) {
         try {
-            byte[] image = recipeService.getImage(new ObjectId(id));
-            HttpHeaders headers = new HttpHeaders();
-            headers.setContentType(MediaType.IMAGE_JPEG);
-            headers.setContentLength(image.length);
-
-            return new ResponseEntity<>(image, headers, 200);
-
+            var recipe = recipeService.getRecipeById(new ObjectId(id));
+            if(!recipe.getStatus()){
+                if(token == null || !Objects.equals(recipe.getAuthor(), jwtService.extractUsername(token))){
+                    return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                            .header("Error-message", "Forbidden")
+                            .build();
+                }
+            }
+            return ResponseEntity.ok(recipe);
         } catch (RecipeNotFoundException ex) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
                     .header("Error-Message", ex.getMessage())
@@ -81,8 +138,15 @@ public class RecipeController {
     }
 
     @PutMapping("/change-status/{id}")
-    public ResponseEntity<Void> changeStatus(@PathVariable String id) {
+    public ResponseEntity<Void> changeStatus(@PathVariable String id,
+                                             @CookieValue(name = "token", required = false) String token ) {
         try {
+            if(token == null ||
+                    !recipeService.getRecipeById(new ObjectId(id)).getAuthor().equals(jwtService.extractUsername(token))){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                        .header("Error-message", "Forbidden")
+                        .build();
+            }
             recipeService.changeStatus(id);
             return ResponseEntity.ok().build();
         } catch (RecipeNotFoundException ex) {
@@ -92,28 +156,6 @@ public class RecipeController {
         }
     }
 
-//    @GetMapping("/image/{recipeId}")
-//    public ResponseEntity<byte[]> getImage(@PathVariable String recipeId) {
-//        try {
-//            String imagePath = "..\\przepisnik\\src\\main\\resources\\static\\recipejpg\\" + recipeId + ".jpg.";
-//            Path filePath = Paths.get(imagePath);
-//            byte[] imageBytes = Files.readAllBytes(filePath);
-//
-//            HttpHeaders headers = new HttpHeaders();
-//            headers.setContentType(MediaType.IMAGE_JPEG);
-//            headers.setContentLength(imageBytes.length);
-//
-//            return ResponseEntity.ok()
-//                    .headers(headers)
-//                    .body(imageBytes);
-//
-//        } catch (Exception ex) {
-//            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-//                    .header("Error-Message", ex.getMessage())
-//                    .build();
-//        }
-//    }
-
     @PostMapping("/search")
     public ResponseEntity<List<ShortRecipeDto>> search(@RequestBody SearchRequest searchRequest){
         Pageable pageable = PageRequest.of(searchRequest.getPageNumber(),searchRequest.getPageSize());
@@ -122,9 +164,15 @@ public class RecipeController {
     }
 
     @PostMapping("/search/private")
-    public ResponseEntity<List<ShortRecipeDto>> searchPrivate(@RequestBody SearchRequest searchRequest){
+    public ResponseEntity<List<ShortRecipeDto>> searchPrivate(@RequestBody SearchRequest searchRequest,
+                                                              @CookieValue(name = "token", required = false) String token ){
 
         //TO DO sprawdzanie uprawnień czy podany autor w request body to user z tokenu
+        if(token == null || !searchRequest.getAuthor().equals(jwtService.extractUsername(token))){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .header("Error-message", "Forbidden")
+                    .build();
+        }
 
         Pageable pageable = PageRequest.of(searchRequest.getPageNumber(),searchRequest.getPageSize());
         return ResponseEntity.ok(recipeService.search(searchRequest.getAuthor(),
@@ -132,8 +180,16 @@ public class RecipeController {
     }
 
     @DeleteMapping("{recipeId}")
-    public ResponseEntity<Void> deleteRecipe(@PathVariable String recipeId){
+    public ResponseEntity<Void> deleteRecipe(@PathVariable String recipeId,
+                                             @CookieValue(name = "token", required = false) String token ){
         try {
+            if(token == null ||
+                    !recipeService.getRecipeById(new ObjectId(recipeId)).getAuthor()
+                    .equals(jwtService.extractUsername(token))){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                        .header("Error-message", "Forbidden")
+                        .build();
+            }
             recipeService.deleteRecipe(new ObjectId(recipeId));
             return ResponseEntity.ok().build();
         } catch (RecipeNotFoundException ex) {
